requirements.txt
anyio==4.9.0
certifi==2025.1.31
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
python-dotenv==1.1.0
python-telegram-bot==20.3
sniffio==1.3.1
